feat: first implementation
This project offers RSS feeds for different online blogs that doesn't
have them. They are accesible via HTTP using the following paths:

/blog-isabel/feed
  https://marmenormarmayor.es/El-blog-de-Isabel/index.html
/verde/blog/feed
  https://elclickverde.com/blog
/verde/reportajes/feed
  https://elclickverde.com/reportajes

The binary offers the way to set the address to listen via CLI, and is
able to show the logs using RUST_LOG variable as defined in env_logger.